Prefer dynamic PT 101 for telephone-event at clock rate 8000
The WebRTC implementation in Google Chrome <= 26 would reject calls
if there was "a telephone-event payload type less than 101"[1] and
as of 2013-06-03, the Google Mail web UI with the VoIP extension
seems to have a similar signalling bug. Experimenting with the web UI
indicates that telephone-events with clock rate != 8000 are irrelevant,
and only clock rate 8000 matters. Hopefully the same was true in WebRTC
(I can't find a libjingle commit that looks likely to have fixed this).

Meanwhile, many SIP implementations and at least one Jingle
implementation (Freeswitch's mod_dingaling) either hard-code payload
type 101 to be telephone-event, or make the payload type for
telephone-event a configuration option. I can't help thinking this
was not how dynamic payload types were meant to work, but
interoperability is interoperability...

This fixes interop when Empathy 3.8 + telepathy-gabble 0.17.4, on a
system with not many codecs installed) calls the Google Mail web
UI. When the same setup is called by a peer that specifies a different
PT for telephone-event:8000 (the Google Mail web UI uses 126 in its
outgoing calls), the peer's choice of PT takes precedence.

[1] https://code.google.com/p/webrtc/issues/detail?id=1783

https://bugs.freedesktop.org/show_bug.cgi?id=65311